🧪 8BH6HY task: isolate worktree resolution coverage
