+++ .agentplane/tasks/202608080551-8BH6HY/pr/diffstat.txt
@@ -1,6 +1,7 @@
- .../src/cli/run-cli.core.task-advance.test.ts      | 237 +++++++++++++++++++++
+ .../src/cli/run-cli.core.task-advance.test.ts      |  61 +++++
+ ...i.core.task-advance.worktree-resolution.test.ts | 292 +++++++++++++++++++++
  .../external-agent-implementation-authority.ts     |   3 +-
  .../commands/task/external-agent-purpose.test.ts   |  19 ++
  .../src/commands/task/external-agent-purpose.ts    |  11 +
- .../src/commands/task/external-agent-supervisor.ts |  30 ++-
- 5 files changed, 283 insertions(+), 17 deletions(-)
+ .../src/commands/task/external-agent-supervisor.ts |  30 +--
+ 6 files changed, 399 insertions(+), 17 deletions(-)
