+++ .agentplane/tasks/202608080551-8BH6HY/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608080551-8BH6HY/accept-external-task-worktree-resolution-results",
   "created_at": "2026-08-08T05:52:09.195Z",
-  "diffstat_sha256": "sha256:4f2a3b697755cdfdf1484138fa6640bd6d8d7d9329a7dab456b9d969e576ff26",
+  "diffstat_sha256": "sha256:e73ceae306261e26d6bca28303bfae661bfddf84898a767cc28e0e0fba305a8c",
   "last_verified_at": "2026-08-08T06:18:57.190Z",
   "last_verified_diffstat_sha256": "sha256:4f2a3b697755cdfdf1484138fa6640bd6d8d7d9329a7dab456b9d969e576ff26",
   "pr_number": 4802,
